fix: solve the methods application issue
To apply method calls of the main CSSStyleSheet to all adopters we need them to have style elements with CSSStyleSheet object attached. It means that we cannot work with disconnected elements, and the idea to work from inside the setter of the adoptedStyleSheets setter does not fit.

That's why we need to track the already connected elements via the MutationObserver and NodeIterator (to catch nested nodes).